requirements.txt
mdutils==1.6.0
PyMuPDF==1.24.2
groq==0.5.0
python-dotenv==1.0.1
black==24.4.0